refactor: rename ReferencesProvider
