+++ packages/providers/src/index.tsx
@@ -1,12 +1,11 @@
 export * from './hooks.js';
 export * from './theme.js';
 export * from './grid.js';
-export * from './references.js';
+export * from './article.js';
 export * from './baseurl.js';
 export * from './ui.js';
 export * from './site.js';
 export * from './search.js';
-export * from './pageKind.js';
 export * from './tabs.js';
 export * from './xref.js';
 export * from './renderers.js';
