+++ themes/article/app/components/Article.tsx
@@ -19,11 +19,10 @@ import {
 } from '@myst-theme/jupyter';
 import { FrontmatterBlock } from '@myst-theme/frontmatter';
 import {
-  ReferencesProvider,
+  ArticleProvider,
   useThemeTop,
   useMediaQuery,
   useProjectManifest,
-  PageKindProvider,
 } from '@myst-theme/providers';
 import type { GenericParent } from 'myst-common';
 import { copyNode } from 'myst-common';
@@ -57,51 +56,50 @@ export function Article({
   const { thebe } = manifest as any;
   const { location } = article;
   return (
-    <ReferencesProvider
+    <ArticleProvider
+      kind={article.kind}
       references={{ ...article.references, article: article.mdast }}
       frontmatter={article.frontmatter}
     >
       <BusyScopeProvider>
         <ExecuteScopeProvider enable={compute?.enabled ?? false} contents={article}>
-          <PageKindProvider pageKind={article.kind}>
-            {!hideTitle && (
-              <FrontmatterBlock
-                frontmatter={{ title, subtitle }}
-                thebe={thebe}
-                location={location}
-                className="mb-5"
-              />
-            )}
-            {!hideOutline && (
-              <div
-                className="block my-10 lg:sticky lg:top-0 lg:z-10 lg:h-0 lg:pt-0 lg:my-0 lg:ml-10 lg:col-margin-right"
-                style={{ top: top + TOP_OFFSET }}
+          {!hideTitle && (
+            <FrontmatterBlock
+              frontmatter={{ title, subtitle }}
+              thebe={thebe}
+              location={location}
+              className="mb-5"
+            />
+          )}
+          {!hideOutline && (
+            <div
+              className="block my-10 lg:sticky lg:top-0 lg:z-10 lg:h-0 lg:pt-0 lg:my-0 lg:ml-10 lg:col-margin-right"
+              style={{ top: top + TOP_OFFSET }}
+            >
+              <DocumentOutline
+                className="relative pt-[2px]"
+                maxdepth={outlineMaxDepth}
+                isMargin={isOutlineMargin}
+                title="In this article"
               >
-                <DocumentOutline
-                  className="relative pt-[2px]"
-                  maxdepth={outlineMaxDepth}
-                  isMargin={isOutlineMargin}
-                  title="In this article"
-                >
-                  <SupportingDocuments />
-                </DocumentOutline>
-              </div>
-            )}
+                <SupportingDocuments />
+              </DocumentOutline>
+            </div>
+          )}
 
-            {compute?.enabled &&
-              compute?.features.notebookCompute &&
-              article.kind === SourceFileKind.Notebook && <NotebookToolbar showLaunch />}
-            <ErrorTray pageSlug={article.slug} />
-            <div id="skip-to-article" />
-            <FrontmatterParts parts={parts} keywords={keywords} hideKeywords={hideKeywords} />
-            <MyST ast={tree.children as GenericParent[]} />
-            <BackmatterParts parts={parts} />
-            <Footnotes />
-            <Bibliography />
-            <ConnectionStatusTray />
-          </PageKindProvider>
+          {compute?.enabled &&
+            compute?.features.notebookCompute &&
+            article.kind === SourceFileKind.Notebook && <NotebookToolbar showLaunch />}
+          <ErrorTray pageSlug={article.slug} />
+          <div id="skip-to-article" />
+          <FrontmatterParts parts={parts} keywords={keywords} hideKeywords={hideKeywords} />
+          <MyST ast={tree.children as GenericParent[]} />
+          <BackmatterParts parts={parts} />
+          <Footnotes />
+          <Bibliography />
+          <ConnectionStatusTray />
         </ExecuteScopeProvider>
       </BusyScopeProvider>
-    </ReferencesProvider>
+    </ArticleProvider>
   );
 }
